requirements.txt
flask==3.1.0
python-dotenv==1.0.1
tavily-python==0.3.0
openai==1.12.0
requests==2.31.0
networkx==3.2.1
python-dateutil==2.8.2
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.10
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
aiohttp>=3.9.0
uvicorn>=0.24.0
firebase-admin>=6.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.104.0
hypercorn>=0.15.0
asgiref>=3.7.2 